hci: install handler for link key request only if needed
Instead of hardcoding this handler in HciDev code, use the event
registration mechanism. In this way the callback code will only be
compiled in if some client asked to register a callback for this event.

In practice this makes a minimal difference, since the callback code is
extremely small, but this sets an example for other events, which might
have more complex handlers.