hci: add a userdata parameter to all methods
Having a single userdata aprameter for all functions can be
inconvenient; this is especially true when we want to call these methods
from an upper layer within the library: we don't want to overwrite the
client's "userdata" property in these cases.

Another solution would be using an additional client, but that risks
complicating things.